maybe fixed /feed? need to first make new user database
